Update tutorial to reflect removal of register_class from API.
